Proposal
Proposal in a Word document. Note: this project was professional bound at FedEx. Because of this, special considerations had to be made to allow for insertions of the external documents. For example, the ERD is a Visio document and therefore one page had to be reserved (and later removed) to ensure the page numbers would align. This is also true for the SQL.