Proposal (including Reference tables)
This section of the final project includes the Proposal. As required, included is the Conceptual level Reference Tables and sources. Whenever reference tables had not already published we were instructed to "create" them. For example, The Best Co. RFT 19 was create for the database need for an Income Table. Note: the proposal was professional bound by an office supply shop. The project proposal is in a Word document. Special considerations were made for the external files that were necessary for inclusion. For example, the ERD is an external Visio document and the SQL was also external. Therefore, pages were reserved (and later removed) to ensure that page numbers were aligned.